src/输入输出篇/接口服务/RPC服务/code/GRpc接口服务/C0/requirements.txt
autopep8==1.4.3
coverage==4.5.3
grpcio==1.19.0
lxml==4.3.3
mypy==0.670
mypy-extensions==0.4.1
pep8==1.7.1
protobuf==3.7.1
pycodestyle==2.5.0
pydocstyle==3.0.0
six==1.12.0
snowballstemmer==1.2.1
typed-ast==1.3.1
voluptuous==0.11.5
aiogrpc